Apply the previous Chess related changes to Checkers.
<h1>Rationale:</h1>

The fix for the ChessFragment illegal argument exception could apply equally to the CheckersFragment class. In so doing the opportunity to coalesce code is overwhelming. This commit represents the capitulation to temptation.

<h1>File changes:</h1>

modified:   app/src/main/java/com/pajato/android/gamechat/exp/Checkerboard.java

- getCellView(): add overloads to handle both Chess and Checkers sharing as much code as possible.
- handleTileBackground(): procedural abstraction to handle the cell background.
- containsPrimaryPiece(), containsSecondayPiece(): checkers helper methods.

modified:   app/src/main/java/com/pajato/android/gamechat/exp/fragment/CheckersFragment.java

- isEvenEvenOrOddOdd(): removed as it became obsolete with the shared checkerboard code.
- containsSecondaryPiece(), containsPrimaryPiece(): moved to checkerboard class.
- makeBoardButton(): morphed into common code on the checkerboard class.
- startGame(): use the new checkerboard functionality.

modified:   app/src/main/java/com/pajato/android/gamechat/exp/fragment/ChessFragment.java

- handleTileBackground(): move to checkerboard class.
- makeBoardButton(): morphed into common code on the checkerboard class.
- startGame(): use the new checkerboard functionality.

modified:   build.gradle

- Summary: new AS update.
